fix coin manager

+++ RagnarokBotWeb/Application/Models/TrapLog.cs
@@ -6,8 +6,8 @@
         public int ScumId { get; set; }
         public string SteamId { get; set; }
         public string TrapName { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
     }
 }
